refactor(web): name test files after the modules they test
params/user/tests.rs named nothing at all, and organization/user_role_tests.rs
named a user_role.rs that does not exist in that directory. Both now follow the
mod_tests.rs / <source>_<feature>_tests.rs shape used elsewhere.
